Refactor models and controllers modules. To ease adding other social networks
