+++ devsocial/controllers/social_network.py
@@ -1,4 +1,4 @@
-# pylint disable:too-few-public-methods
+# pylint: disable:too-few-public-methods
 # pylint: disable=missing-class-docstring
 # pylint: disable=missing-function-docstring
 """
@@ -6,7 +6,7 @@ Each social network will have a the business logic implemented in this module.
 Note: Please implement the base class methods listed below if you want to create you own social network controller.
 """
 from abc import abstractmethod
-from devsocial.models.developer import GitHubDeveloper, BaseDeveloper, TwitterDeveloper
+from devsocial.models.developer import BaseDeveloper
 
 
 # pylint: disable=too-few-public-methods
@@ -17,23 +17,3 @@ class BaseDevSocialNet:
     @abstractmethod
     def connected(self, developer1: BaseDeveloper, developer2: BaseDeveloper):
         pass
-
-
-# pylint: disable=too-few-public-methods, useless-super-delegation
-class GitHubDevSocialNet(BaseDevSocialNet):
-    def __init__(self):
-        super().__init__()
-
-    def connected(self, developer1: GitHubDeveloper, developer2: GitHubDeveloper):
-        if developer1 != developer2:
-            return bool(set(developer1.organizations) & set(developer2.organizations))
-        return False
-
-
-# pylint: disable=too-few-public-methods, useless-super-delegation
-class TwitterDevSocialNet(BaseDevSocialNet):
-    def __init__(self):
-        super().__init__()
-
-    def connected(self, developer1: TwitterDeveloper, developer2: TwitterDeveloper):
-        return developer1 in developer2.followers and developer2 in developer1.followers
