+++ tests/test_github.py
@@ -1,7 +1,18 @@
 # pylint: disable=missing-module-docstring, missing-function-docstring, unused-variable
-from devsocial.controllers.social_network import GitHubDevSocialNet, TwitterDevSocialNet
-from devsocial.models.developer import GitHubDeveloper, TwitterDeveloper
 from devsocial.models.organization import Organization
+from devsocial.github.models import GitHubDeveloper
+from devsocial.github.controllers import GitHubDevSocialNet
+
+
+def test_github_dev_belongs_organization():
+    organizations = [Organization(name) for name in ("Nuclear Plant", "Moe's Tavern")]
+    github_dev: GitHubDeveloper = GitHubDeveloper('homer', organizations=organizations)
+    assert organizations[0] in github_dev.organizations
+
+
+def test_github_dev_belongs_no_organization():
+    github_dev: GitHubDeveloper = GitHubDeveloper('homer')
+    assert not github_dev.organizations
 
 
 def test_two_developers_are_github_connected():
@@ -19,29 +30,6 @@ def test_two_developers_are_not_github_connected():
     github_dev2: GitHubDeveloper = GitHubDeveloper('krasty', organizations=dev2_organizations)
     github_net: GitHubDevSocialNet = GitHubDevSocialNet()
     assert not github_net.connected(github_dev1, github_dev2)
-
-
-def test_two_developers_are_twitter_connected():
-    twitter_dev1: TwitterDeveloper = TwitterDeveloper('homer')
-    twitter_dev2: TwitterDeveloper = TwitterDeveloper('mrburns')
-    twitter_dev1.followers.append(twitter_dev2)
-    twitter_dev2.followers.append(twitter_dev1)
-    twitter_net: TwitterDevSocialNet = TwitterDevSocialNet()
-    assert twitter_net.connected(twitter_dev1, twitter_dev2)
-
-
-def test_two_developers_are_not_twitter_connected():
-    twitter_dev1: TwitterDeveloper = TwitterDeveloper('homer')
-    twitter_dev2: TwitterDeveloper = TwitterDeveloper('flanders')
-    twitter_dev2.followers.append(twitter_dev1)
-    twitter_net: TwitterDevSocialNet = TwitterDevSocialNet()
-    assert not twitter_net.connected(twitter_dev1, twitter_dev2)
-
-
-def test_check_same_developer_twitter_connected():
-    twitter_dev1: TwitterDeveloper = TwitterDeveloper('homer')
-    twitter_net: TwitterDevSocialNet = TwitterDevSocialNet()
-    assert not twitter_net.connected(twitter_dev1, twitter_dev1)
 
 
 def test_check_same_developer_github_connected():
