+++ tests/test_models.py
@@ -1,7 +1,6 @@
 # pylint: disable=missing-module-docstring, missing-function-docstring, unused-variable
 import pytest
-from devsocial.models.developer import BaseDeveloper, GitHubDeveloper, TwitterDeveloper, HandleType
-from devsocial.models.organization import Organization
+from devsocial.models.developer import BaseDeveloper
 
 
 def test_create_developer():
@@ -28,25 +27,3 @@ def test_developers_not_equal():
 
     with pytest.raises(AssertionError):
         assert dev1.handle == dev2.handle
-
-
-def test_github_dev_belongs_organization():
-    organizations = [Organization(name) for name in ("Nuclear Plant", "Moe's Tavern")]
-    github_dev: GitHubDeveloper = GitHubDeveloper('homer', organizations=organizations)
-    assert organizations[0] in github_dev.organizations
-
-
-def test_github_dev_belongs_no_organization():
-    github_dev: GitHubDeveloper = GitHubDeveloper('homer')
-    assert not github_dev.organizations
-
-
-def test_twitter_dev_has_followers():
-    followers = ([HandleType(handle) for handle in ('homer', 'moe')])
-    twitter_dev: TwitterDeveloper = TwitterDeveloper('krasty', followers=followers)
-    assert followers[0] in twitter_dev.followers
-
-
-def test_twitter_dev_has_no_followers():
-    twitter_dev: TwitterDeveloper = TwitterDeveloper('krasty')
-    assert not twitter_dev.followers
